License added.
This is currently the University of Illinois/NCSA Open Source License, a lax
permissive license synthesizing the MIT/X11 and 3-clause BSD licenses. While we
typically prefer the GPLv3, copyleft would seem to be inappropriate for dotfiles
intended to be unprofitably hijacked at will by others. Corpulent laxity it is!