Ebuild and Python comment leaders fixed.
The "#" character prefixing comments in such languages is now unconditionally
parsed as a comment leader, regardless of whether such character is immediately
followed by whitespace or not. This appears to be an outstanding bug in both the
"python-mode" plugin and Gentoo's ebuild plugin. (Such are our trials and minute
tribulations.)