Preferred shell coerced to "bash" if available.
Vim's preferred shell is now coerced to the absolute path of the "bash" command
if such command is in the current $PATH. Since Vim offers a builtin for testing
whether a command is in the current $PATH (i.e., executable()) but *NOT* for
getting the absolute path of such command, this was surprisingly troublesome.

Begone, foul insufficiencies!