Large file syntax highlighting kludge.
Large file syntax highlighting appears to be fundamentally broken as of
Vim >= 8.0.0647. As a temporary solution, we adopt the horrible (but
unavoidable) kludge recommended by Vim developers of permanently
increasing the "redrawtime" (i.e., maximum number of milliseconds Vim
permits itself to syntax highlight large buffers in) from the
insufficient default of 2,000 milliseconds to the substantially larger
(but equally arbitrary) value of 10,000 milliseconds.

Yes, this is atrocious. For further details, see ongoing discussion at:
     https://github.com/vim/vim/issues/2790